Fixes following Ethermine stats module upgrades
A few bits broke following the recent changes to the Ethermine API.  This patch
fixes those issues.  Also switched to distributing config.php.dist, so as not
to accidentally overwrite your config file.  There will be a few new views i'll
add soon, thanks to the improvements in the Ethermine API.